# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RuiQi 应用防火墙管理系统 API",
        "title": "RuiQi-WAF API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/HUAHUAI23/simple-waf",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:2333",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/ip-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有IP组列表，支持分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP组管理"
                ],
                "summary": "获取IP组列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取IP组列表成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.IPGroupListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的IP地址组，用于后续IP规则匹配",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP组管理"
                ],
                "summary": "创建IP组",
                "parameters": [
                    {
                        "description": "IP组信息",
                        "name": "ipGroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IPGroupCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IP组创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.IPGroup"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "409": {
                        "description": "IP组名称已存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/ip-groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取IP组详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP组管理"
                ],
                "summary": "获取单个IP组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取IP组详情成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.IPGroup"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的ID格式",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "IP组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定IP组的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP组管理"
                ],
                "summary": "更新IP组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "IP组更新信息",
                        "name": "ipGroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IPGroupUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IP组更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.IPGroup"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止操作系统默认IP组",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "IP组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "409": {
                        "description": "IP组名称已存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定的IP组，系统默认IP组不允许删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP组管理"
                ],
                "summary": "删除IP组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IP组删除成功",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponseNoData"
                        }
                    },
                    "400": {
                        "description": "无效的ID格式",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止删除系统默认IP组",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "IP组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.IPGroupCreateRequest": {
            "description": "创建IP组的请求参数",
            "type": "object",
            "required": [
                "items",
                "name"
            ],
            "properties": {
                "items": {
                    "description": "IP地址或CIDR列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"192.168.1.1\"]"
                    ]
                },
                "name": {
                    "description": "IP组名称",
                    "type": "string",
                    "example": "内部服务器"
                }
            }
        },
        "dto.IPGroupListResponse": {
            "description": "IP组列表响应",
            "type": "object",
            "properties": {
                "items": {
                    "description": "IP组列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.IPGroup"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "dto.IPGroupUpdateRequest": {
            "description": "更新IP组的请求参数",
            "type": "object",
            "properties": {
                "items": {
                    "description": "IP地址或CIDR列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"192.168.1.1\"]"
                    ]
                },
                "name": {
                    "description": "IP组名称",
                    "type": "string",
                    "example": "内部服务器"
                }
            }
        },
...
        "model.ErrResponse": {
            "description": "错误的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "参数错误"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.ErrResponseDontShowError": {
            "description": "错误的API响应标准格式,不展示 error",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.IPGroup": {
            "description": "IP地址组信息，包含组名和IP地址列表",
            "type": "object",
            "properties": {
                "id": {
                    "description": "组唯一标识符",
                    "type": "string",
                    "example": "60d21b4367d0d8992e89e964"
                },
                "items": {
                    "description": "IP地址或CIDR列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['192.168.1.1'",
                        " '10.0.0.1/24']"
                    ]
                },
                "name": {
                    "description": "组名称",
                    "type": "string",
                    "example": "内部服务器"
                }
            }
        },
...
        "model.SuccessResponse": {
            "description": "成功的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.SuccessResponseNoData": {
            "description": "成功的API响应标准格式,没有 data",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer {token} 格式进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}